ml-service/requirements.txt
# Essential ML Service Requirements
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
Pillow==10.1.0
numpy==1.24.3
python-dotenv==1.0.0
pydantic==2.5.0
aiofiles==23.2.1
httpx==0.25.2

# Optional: Lightweight ML (commented out heavy dependencies for now)
# opencv-python==4.8.1.78
# tensorflow==2.15.0
# scikit-learn==1.3.2
# pandas==2.1.4
# torch==2.1.1
# torchvision==0.16.1
# transformers==4.36.0
# datasets==2.14.7

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1
